+++ data-raw/devstuffs.R
@@ -31,7 +31,7 @@ my_desc$set_authors(c(
 # Remove some author fields
 my_desc$del("Maintainer")
 # Set the version
-my_desc$set_version("3.0.0.9001")
+my_desc$set_version("3.0.0.9002")
 # The title of your package
 my_desc$set(Title = "MIRACUM DQA Tool")
 # The description of your package