+++ inst/application/_utilities/MDR/plausibilities.py
@@ -191,7 +191,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item08",
               "description": "Nur bei männlichen Patienten sind sonstige angeborene Fehlbildungen der männlichen Genitalorgane (ICD Q55*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:Q55",
+                "i2b2": "^(ICD10\\:)Q55",
                 "fhir_gw": "^Q55"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -207,7 +207,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item09",
               "description": "Nur bei männlichen Patienten ist Nondescensus testis (ICD Q53*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:Q53",
+                "i2b2": "^(ICD10\\:)Q53",
                 "fhir_gw": "^Q53"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -223,7 +223,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item10",
               "description": "Nur bei männlichen Patienten ist Hypospadie (ICD Q54*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:Q54",
+                "i2b2": "^(ICD10\\:)Q54",
                 "fhir_gw": "^Q54"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -239,7 +239,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item11",
               "description": "Nur bei weiblichen Patientinnen sind angeborene Fehlbildungen der Ovarien, der Tubae uterinae und der Ligg. lata uteri (ICD Q50*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:Q50",
+                "i2b2": "^(ICD10\\:)Q50",
                 "fhir_gw": "^Q50"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -255,7 +255,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item12",
               "description": "Nur bei weiblichen Patientinnen sind angeborene Fehlbildungen des Uterus und der Cervix uteri (ICD Q51*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:Q51",
+                "i2b2": "^(ICD10\\:)Q51",
                 "fhir_gw": "^Q51"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -271,7 +271,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item13",
               "description": "Nur bei weiblichen Patientinnen sind sonstige angeborene Fehlbildungen der weiblichen Genitalorgane (ICD Q52*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:Q52",
+                "i2b2": "^(ICD10\\:)Q52",
                 "fhir_gw": "^Q52"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -287,7 +287,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item14",
               "description": "Nur bei weiblichen Patientinnen sind abnorme Befunde in Untersuchungsmaterialien aus den weiblichen Genitalorganen (ICD R87*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:R87",
+                "i2b2": "^(ICD10\\:)R87",
                 "fhir_gw": "^R87"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -303,7 +303,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item15",
               "description": "Nur bei männlichen Patienten sind abnorme Befunde in Untersuchungsmaterialien aus den männlichen Genitalorganen (ICD R86*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:R86",
+                "i2b2": "^(ICD10\\:)R86",
                 "fhir_gw": "^R86"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -319,7 +319,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item16",
               "description": "Nur bei männlichen Patienten sind gutartige Neubildung der männlichen Genitalorgane (ICD D29*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:D29",
+                "i2b2": "^(ICD10\\:)D29",
                 "fhir_gw": "^D29"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -335,7 +335,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item17",
               "description": "Nur bei weiblichen Patientinnen sind gutartige Neubildung der Brustdrüse [Mamma] (ICD D24*), Leiomyom des Uterus (ICD D25*), sonstige gutartige Neubildungen des Uterus (ICD D26*), gutartige Neubildung des Ovars (ICD D27*) und gutartige Neubildung sonstiger und nicht näher bezeichneter weiblicher Genitalorgane (ICD D28*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:D2[4-8]",
+                "i2b2": "^(ICD10\\:)D2[4-8]",
                 "fhir_gw": "^D2[4-8]"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -351,7 +351,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item18",
               "description": "Nur bei weiblichen Patientinnen sind Neubildung unsicheren oder unbekannten Verhaltens der weiblichen Genitalorgane (ICD D39*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:D39",
+                "i2b2": "^(ICD10\\:)D39",
                 "fhir_gw": "^D39"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -367,7 +367,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item19",
               "description": "Nur bei männlichen Patienten sind Neubildung unsicheren oder unbekannten Verhaltens der männlichen Genitalorgane (ICD D40*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:D40",
+                "i2b2": "^(ICD10\\:)D40",
                 "fhir_gw": "^D40"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -383,7 +383,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item20",
               "description": "Nur bei männlichen Patienten sind Carcinoma in situ sonstiger und nicht näher bezeichneter Genitalorgane: Penis, Prostata, sonstige und nicht näher bezeichnete männliche Genitalorgane (ICD D07.4-6) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:D07\\.[4-6]",
+                "i2b2": "^(ICD10\\:)D07\\.[4-6]",
                 "fhir_gw": "^D07\\.[4-6]"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -399,7 +399,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item21",
               "description": "Nur bei weiblichen Patientinnen sind Carcinoma in situ sonstiger und nicht näher bezeichneter Genitalorgane: Endometrium, Vulva, Vagina, sonstige und nicht näher bezeichnete weibliche Genitalorgane (ICD D07.0-3) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:D07\\.[0-3]",
+                "i2b2": "^(ICD10\\:)D07\\.[0-3]",
                 "fhir_gw": "^D07\\.[0-3]"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -415,7 +415,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item22",
               "description": "Nur bei weiblichen Patientinnen ist Tetanus während der Schwangerschaft, der Geburt und des Wochenbettes (ICD A34) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:A34",
+                "i2b2": "^(ICD10\\:)A34",
                 "fhir_gw": "^A34"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -431,7 +431,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item23",
               "description": "Nur bei weiblichen Patientinnen sind Psychische oder Verhaltensstörungen im Wochenbett, anderenorts nicht klassifiziert (ICD F53) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:F53",
+                "i2b2": "^(ICD10\\:)F53",
                 "fhir_gw": "^F53"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -447,7 +447,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item24",
               "description": "Nur bei weiblichen Patientinnen ist Osteomalazie im Wochenbett (ICD M83.0*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:M83\\.0",
+                "i2b2": "^(ICD10\\:)M83\\.0",
                 "fhir_gw": "^M83\\.0"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",
@@ -463,7 +463,7 @@ class WritePlausibilities(MDRHandling):
               "name": "Pl.atemporal.Item25",
               "description": "Nur bei männlichen Patienten ist Testikuläre Dysfunktion (ICD E29*) als Krankenhausdiagnose erlaubt.",
               "filter": {
-                "i2b2": "^(ICD10\\:E29",
+                "i2b2": "^(ICD10\\:)E29",
                 "fhir_gw": "^E29"
               },
               "join_crit": "Fall.Versorgungsstellenkontakt.Aufnahmenummer",